requirements.txt
requests
shapely
scikit-image
imgaug
pyclipper
lmdb
tqdm
numpy<2.0
rapidfuzz
opencv-python
opencv-contrib-python
cython
Pillow
pyyaml
requests
albumentations==1.4.10
